refactor for scsi support node code
