--- node/pkg/driver/device_connectivity/device_connectivity_fc_helper_test.go
@@ -1,165 +1,0 @@
-/**
- * Copyright 2019 IBM Corp.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package device_connectivity_test
-
-import (
-	"fmt"
-	"reflect"
-	"testing"
-
-	gomock "github.com/golang/mock/gomock"
-	mocks "github.com/ibm/ibm-block-csi-driver/node/mocks"
-	"github.com/ibm/ibm-block-csi-driver/node/pkg/driver/device_connectivity"
-)
-
-func TestHelperGetFcHostIDs(t *testing.T) {
-	testCases := []struct {
-		name                  string
-		ioutilReadFileReturns []ioutilReadFileReturn
-		arrayIdentifier       string
-
-		expErrType        reflect.Type
-		expErr            error
-		expHostList       []int
-		globReturnMatches []string
-		globReturnErr     error
-	}{
-		{
-			name:              "Should fail when FilepathGlob return error",
-			arrayIdentifier:   "fakeWWN",
-			globReturnMatches: nil,
-			globReturnErr:     fmt.Errorf("error"),
-			expErr:            fmt.Errorf("error"),
-			expHostList:       nil,
-		},
-		{
-			name:              "Should fail when FilepathGlob return without any hosts target files at all",
-			arrayIdentifier:   "fakeWWN",
-			globReturnMatches: nil,
-			globReturnErr:     nil,
-
-			expErrType:  reflect.TypeOf(&device_connectivity.ConnectivityFcHostTargetNotFoundError{}),
-			expHostList: nil,
-		},
-		{
-			name: "Should fail when all port state is offline",
-			ioutilReadFileReturns: []ioutilReadFileReturn{
-				ioutilReadFileReturn{
-					ReadFileParam: "/sys/class/fc_host/host33/port_state",
-					data:          []byte("offline"),
-					err:           nil,
-				},
-				ioutilReadFileReturn{
-					ReadFileParam: "/sys/class/fc_host/host34/port_state",
-					data:          []byte("offline"),
-					err:           nil,
-				},
-			},
-			arrayIdentifier: "fakeWWN",
-			globReturnMatches: []string{
-				"/sys/class/fc_host/host33/port_state",
-				"/sys/class/fc_host/host34/port_state",
-			},
-			globReturnErr: nil,
-
-			expErrType:  reflect.TypeOf(&device_connectivity.ConnectivityFcHostTargetNotFoundError{}),
-			expHostList: nil,
-		},
-
-		{
-			name: "Should succeed to find host33 and host34(host 35 offline, hott36 return error)",
-			ioutilReadFileReturns: []ioutilReadFileReturn{
-				ioutilReadFileReturn{
-					ReadFileParam: "/sys/class/fc_host/host33/port_state",
-					data:          []byte("online"),
-					err:           nil,
-				},
-				ioutilReadFileReturn{
-					ReadFileParam: "/sys/class/fc_host/host34/port_state",
-					data:          []byte("Online"),
-					err:           nil,
-				},
-				ioutilReadFileReturn{
-					ReadFileParam: "/sys/class/fc_host/host35/port_state",
-					data:          []byte("Offline"),
-					err:           nil,
-				},
-				ioutilReadFileReturn{
-					ReadFileParam: "/sys/class/fc_host/host36/port_state",
-					data:          nil,
-					err:           fmt.Errorf("error"),
-				},
-			},
-			arrayIdentifier: "fakeWWN",
-
-			globReturnMatches: []string{
-				"/sys/class/fc_host/host33/port_state",
-				"/sys/class/fc_host/host34/port_state",
-				"/sys/class/fc_host/host35/port_state",
-				"/sys/class/fc_host/host36/port_state",
-			},
-			globReturnErr: nil,
-
-			expErrType:  nil,
-			expHostList: []int{33, 34},
-		},
-	}
-
-	for _, tc := range testCases {
-
-		t.Run(tc.name, func(t *testing.T) {
-			mockCtrl := gomock.NewController(t)
-			defer mockCtrl.Finish()
-
-			fake_executer := mocks.NewMockExecuterInterface(mockCtrl)
-
-			fake_executer.EXPECT().FilepathGlob(device_connectivity.FC_HOST_SYSFS_PATH+"/host*/port_state").Return(tc.globReturnMatches, tc.globReturnErr)
-
-			var mcalls []*gomock.Call
-			for _, r := range tc.ioutilReadFileReturns {
-				call := fake_executer.EXPECT().IoutilReadFile(r.ReadFileParam).Return(r.data, r.err)
-				mcalls = append(mcalls, call)
-			}
-			gomock.InOrder(mcalls...)
-
-			helperFc := device_connectivity.NewOsDeviceConnectivityHelperFc(fake_executer)
-
-			returnHostList, err := helperFc.GetFcHostIDs()
-			if tc.expErr != nil || tc.expErrType != nil {
-				if err == nil {
-					t.Fatalf("Expected to fail with error, got success.")
-				}
-				if tc.expErrType != nil {
-					if reflect.TypeOf(err) != tc.expErrType {
-						t.Fatalf("Expected error type %v, got different error %v", tc.expErrType, reflect.TypeOf(err))
-					}
-				} else {
-					if err.Error() != tc.expErr.Error() {
-						t.Fatalf("Expected error code %s, got %s", tc.expErr, err.Error())
-					}
-				}
-			}
-
-			if len(tc.expHostList) == 0 && len(returnHostList) == 0 {
-				return
-			} else if !reflect.DeepEqual(returnHostList, tc.expHostList) {
-				t.Fatalf("Expected found hosts dirs %v, got %v", tc.expHostList, returnHostList)
-			}
-
-		})
-	}
-}
