--- node/pkg/driver/device_connectivity/device_connectivity_iscsi_helper_test.go
@@ -1,190 +1,0 @@
-/**
- * Copyright 2019 IBM Corp.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package device_connectivity_test
-
-import (
-	"fmt"
-	"reflect"
-	"testing"
-
-	gomock "github.com/golang/mock/gomock"
-	mocks "github.com/ibm/ibm-block-csi-driver/node/mocks"
-	"github.com/ibm/ibm-block-csi-driver/node/pkg/driver/device_connectivity"
-)
-
-type ioutilReadFileReturn struct {
-	ReadFileParam string // The param that the IoutilReadDir recive on each call.
-	data          []byte
-	err           error
-}
-
-func TestHelperGetIscsiSessionHostsForArrayIQN(t *testing.T) {
-	testCases := []struct {
-		name                  string
-		ioutilReadFileReturns []ioutilReadFileReturn
-		arrayIdentifier       string
-
-		expErrType        reflect.Type
-		expErr            error
-		expHostList       []int
-		globReturnMatches []string
-		globReturnErr     error
-	}{
-		{
-			name:              "Should fail when FilepathGlob return error",
-			arrayIdentifier:   "fakeIQN",
-			globReturnMatches: nil,
-			globReturnErr:     fmt.Errorf("error"),
-			expErr:            fmt.Errorf("error"),
-			expHostList:       nil,
-		},
-		{
-			name:              "Should fail when FilepathGlob return without any hosts target files at all",
-			arrayIdentifier:   "fakeIQN",
-			globReturnMatches: nil,
-			globReturnErr:     nil,
-
-			expErrType:  reflect.TypeOf(&device_connectivity.ConnectivityIscsiStorageTargetNotFoundError{}),
-			expHostList: nil,
-		},
-		{
-			name: "Should fail when array IQN was not found in target files at all",
-			ioutilReadFileReturns: []ioutilReadFileReturn{
-				ioutilReadFileReturn{
-					ReadFileParam: "/sys/class/iscsi_host/host1/device/session1/iscsi_session/session1/targetname",
-					data:          []byte("fakeIQN_OTHER"),
-					err:           nil,
-				},
-				ioutilReadFileReturn{
-					ReadFileParam: "/sys/class/iscsi_host/host2/device/session2/iscsi_session/session2/targetname",
-					data:          []byte("fakeIQN_OTHER"),
-					err:           nil,
-				},
-			},
-			arrayIdentifier: "fakeIQN",
-			globReturnMatches: []string{
-				"/sys/class/iscsi_host/host1/device/session1/iscsi_session/session1/targetname",
-				"/sys/class/iscsi_host/host2/device/session2/iscsi_session/session2/targetname",
-			},
-			globReturnErr: nil,
-
-			expErrType:  reflect.TypeOf(&device_connectivity.ConnectivityIscsiStorageTargetNotFoundError{}),
-			expHostList: nil,
-		},
-		{
-			name: "Should fail when array IQN found but hostX where X is not int",
-			ioutilReadFileReturns: []ioutilReadFileReturn{
-				ioutilReadFileReturn{
-					ReadFileParam: "/sys/class/iscsi_host/hostX/device/session1/iscsi_session/session1/targetname",
-					data:          []byte("fakeIQN"),
-					err:           nil,
-				},
-			},
-			arrayIdentifier: "fakeIQN",
-
-			globReturnMatches: []string{
-				"/sys/class/iscsi_host/hostX/device/session1/iscsi_session/session1/targetname",
-			},
-			globReturnErr: nil,
-
-			expErrType:  reflect.TypeOf(&device_connectivity.ConnectivityIscsiStorageTargetNotFoundError{}),
-			expHostList: nil,
-		},
-
-		{
-			name: "Should succeed to find host1 and host2 for the array IQN (while host3 is not from this IQN and also host666 fail ignore)",
-			ioutilReadFileReturns: []ioutilReadFileReturn{
-				ioutilReadFileReturn{
-					ReadFileParam: "/sys/class/iscsi_host/host1/device/session1/iscsi_session/session1/targetname",
-					data:          []byte("fakeIQN"),
-					err:           nil,
-				},
-				ioutilReadFileReturn{
-					ReadFileParam: "/sys/class/iscsi_host/host2/device/session1/iscsi_session/session1/targetname",
-					data:          []byte("fakeIQN"),
-					err:           nil,
-				},
-				ioutilReadFileReturn{
-					ReadFileParam: "/sys/class/iscsi_host/host3/device/session1/iscsi_session/session1/targetname",
-					data:          []byte("fakeIQN_OTHER"),
-					err:           nil,
-				},
-				ioutilReadFileReturn{
-					ReadFileParam: "/sys/class/iscsi_host/host666/device/session1/iscsi_session/session1/targetname",
-					data:          nil,
-					err:           fmt.Errorf("error"),
-				},
-			},
-			arrayIdentifier: "fakeIQN",
-
-			globReturnMatches: []string{
-				"/sys/class/iscsi_host/host1/device/session1/iscsi_session/session1/targetname",
-				"/sys/class/iscsi_host/host2/device/session1/iscsi_session/session1/targetname",
-				"/sys/class/iscsi_host/host3/device/session1/iscsi_session/session1/targetname",
-				"/sys/class/iscsi_host/host666/device/session1/iscsi_session/session1/targetname",
-			},
-			globReturnErr: nil,
-
-			expErrType:  nil,
-			expHostList: []int{1, 2},
-		},
-	}
-
-	for _, tc := range testCases {
-
-		t.Run(tc.name, func(t *testing.T) {
-			mockCtrl := gomock.NewController(t)
-			defer mockCtrl.Finish()
-
-			fake_executer := mocks.NewMockExecuterInterface(mockCtrl)
-
-			fake_executer.EXPECT().FilepathGlob(device_connectivity.IscsiHostRexExPath).Return(tc.globReturnMatches, tc.globReturnErr)
-
-			var mcalls []*gomock.Call
-			for _, r := range tc.ioutilReadFileReturns {
-				call := fake_executer.EXPECT().IoutilReadFile(r.ReadFileParam).Return(r.data, r.err)
-				mcalls = append(mcalls, call)
-			}
-			gomock.InOrder(mcalls...)
-
-			helperIscsi := device_connectivity.NewOsDeviceConnectivityHelperIscsi(fake_executer)
-
-			returnHostList, err := helperIscsi.GetIscsiSessionHostsForArrayIQN(tc.arrayIdentifier)
-			if tc.expErr != nil || tc.expErrType != nil {
-				if err == nil {
-					t.Fatalf("Expected to fail with error, got success.")
-				}
-				if tc.expErrType != nil {
-					if reflect.TypeOf(err) != tc.expErrType {
-						t.Fatalf("Expected error type %v, got different error %v", tc.expErrType, reflect.TypeOf(err))
-					}
-				} else {
-					if err.Error() != tc.expErr.Error() {
-						t.Fatalf("Expected error code %s, got %s", tc.expErr, err.Error())
-					}
-				}
-			}
-
-			if len(tc.expHostList) == 0 && len(returnHostList) == 0 {
-				return
-			} else if !reflect.DeepEqual(returnHostList, tc.expHostList) {
-				t.Fatalf("Expected found hosts dirs %v, got %v", tc.expHostList, returnHostList)
-			}
-
-		})
-	}
-}
